Moved test-setup to Vitest
