+++ test/vitest/tests/test-setup.ts
@@ -3,15 +3,15 @@ import FsProvider from '../../../src/providers/generic/file/FsProvider';
 import Profile from '../../../src/model/Profile';
 import FileUtils from '../../../src/utils/FileUtils';
 import NodeFs from '../../../src/providers/generic/file/NodeFs';
-import StubFsProvider from '../../vitest/stubs/providers/stub.FsProvider';
-import StubInteractionProvider from '../../vitest/stubs/providers/stub.InteractionProvider';
+import StubFsProvider from '../stubs/providers/stub.FsProvider';
+import StubInteractionProvider from '../stubs/providers/stub.InteractionProvider';
 import InteractionProvider from '../../../src/providers/ror2/system/InteractionProvider';
-import StubLinkProvider from '../../vitest/stubs/providers/stub.LinkProvider';
+import StubLinkProvider from '../stubs/providers/stub.LinkProvider';
 import LinkProvider from '../../../src/providers/components/LinkProvider';
-import StubProfileProvider from '../../vitest/stubs/providers/stub.ProfileProvider';
-import ProfileProvider from 'src/providers/ror2/model_implementation/ProfileProvider';
+import StubProfileProvider from '../stubs/providers/stub.ProfileProvider';
+import ProfileProvider from '../../../src/providers/ror2/model_implementation/ProfileProvider';
 import {providePathImplementation} from "../../../src/providers/node/path/path";
-import {TestPathProvider} from "../../vitest/stubs/providers/node/Node.Path.Provider";
+import {TestPathProvider} from "../stubs/providers/node/Node.Path.Provider";
 
 export default class TestSetup {
